+++ src/components/title1/Title1Map.tsx
@@ -11,7 +11,6 @@ import "../../styles/map.css";
 import DrawLegend from "../shared/DrawLegend";
 import legendConfig from "../../utils/legendConfig.json";
 import { ShortFormat } from "../shared/ConvertionFormats";
-import { compareWithDollarSign } from "../shared/TableCompareFunctions";
 
 const geoUrl = "https://cdn.jsdelivr.net/npm/us-atlas@3/states-10m.json";
 
@@ -323,7 +322,7 @@ const Title1Map = ({
     const searchKey = !subprogram ? program || subtitle : subprogram;
     const customScale = legendConfig[searchKey];
     const colorScale = d3.scaleThreshold(customScale, mapColor);
-    const zeroPoints: string[] = [];
+    const zeroPoints = [];
     statePerformance[year].forEach((state) => {
         if (subtitle && program) {
             const programList = state.programs;
@@ -349,8 +348,6 @@ const Title1Map = ({
                     programData={quantizeArray}
                     prepColor={mapColor}
                     emptyState={zeroPoints}
-                    initRatioLarge={0.6}
-                    initRatioSmall={0.5}
                 />
             </Box>
             <MapChart